Update to http4s 0.18
