+++ .gitignore
@@ -1,5 +1,22 @@
+
+### Scala IDE specific
+.cache-main
+.scala_dependencies
+.worksheet
+### Scala
+*.class
+*.log
+### SBT 
+dist/*
 target/
-boot/
 lib_managed/
 src_managed/
+project/boot/
 project/plugins/project/
+.history
+.cache
+.lib/
+### JetBrains 
+.idea
+*.iws
+/out/
