+++ src/main/g8/default.properties
@@ -2,5 +2,5 @@ name = http4s quickstart
 organization = com.example
 package = $organization$.$name;format="norm,word"$
 
-scala_version = 2.12.2
-http4s_version = 0.15.11a
+scala_version = 2.12.3
+http4s_version = 0.18.0-M1
