--- src/main/g8/src/main/scala/$package__packaged$/HelloWorld.scala
@@ -1,14 +1,0 @@
-package $package$
-
-import io.circe._
-import org.http4s._
-import org.http4s.circe._
-import org.http4s.server._
-import org.http4s.dsl._
-
-object HelloWorld {
-  val service = HttpService {
-    case GET -> Root / "hello" / name =>
-      Ok(Json.obj("message" -> Json.fromString(s"Hello, \${name}")))
-  }
-}
